Use runtime.cgo.Handle instead of a global Context registry
The Go Context object creates a Handle referring to itself and passes that to
the C++ code to remember, instead of an integer index in a registry.

+++ v8go.h
@@ -190,7 +190,7 @@ extern void CPUProfileDelete(CPUProfile* ptr);
 
 extern ContextPtr NewContext(IsolatePtr iso_ptr,
                              TemplatePtr global_template_ptr,
-                             int ref);
+                             uintptr_t ref);
 extern void ContextFree(ContextPtr ptr);
 extern RtnValue RunScript(ContextPtr ctx_ptr,
                           const char* source, int sourceLen,
